pyproject.toml
[project]
name = "veeam-presence"
version = "1.0.0"
description = "Veeam Presence — Office attendance intelligence agent"
requires-python = ">=3.11"
dependencies = [
    # Agent Framework + Azure OpenAI
    "agent-framework[azure]>=1.0.0rc4",
    "azure-identity>=1.15.0",
    "openai>=1.40.0",
    # Data
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "holidays>=0.40",
    # Web
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0",
    # Cards
    "pydantic>=2.0",
]

[project.optional-dependencies]
m365 = [
    "microsoft-agents-hosting-core>=0.8.0",
    "microsoft-agents-hosting-fastapi>=0.8.0",
    "microsoft-agents-authentication-msal>=0.8.0",
    "msal>=1.31.1",
    "PyJWT[crypto]>=2.10.1",
]
pipeline = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "holidays>=0.40",
    "requests>=2.28.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests", "agent/tests", "m365_wrapper/tests", "dev_ui/tests", "infra/tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.uv]
prerelease = "allow"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
